Adds localized metadata and refactors auth UI
