+++ apps/web/src/app/[locale]/(dashboard)/(home)/page.tsx
@@ -1,7 +1,14 @@
+import type { Metadata } from 'next';
+import { getTranslations } from 'next-intl/server';
 import { getAuthFromCookie } from '@/lib/auth';
 import DashboardClient from './dashboard-client';
 import { getDirectories } from '@/app/actions/dashboard/directories';
 import { GET_DIRECTORY_LIST_LIMIT } from '@/lib/constants';
+
+export async function generateMetadata(): Promise<Metadata> {
+    const t = await getTranslations('metadata.pages');
+    return { title: t('dashboard') };
+}
 
 export default async function Dashboard() {
     const user = await getAuthFromCookie();
