+++ apps/web/src/app/[locale]/(dashboard)/directories/[id]/members/page.tsx
@@ -1,7 +1,14 @@
+import type { Metadata } from 'next';
+import { getTranslations } from 'next-intl/server';
 import { directoryAPI, membersAPI } from '@/lib/api';
 import { MembersPage } from '@/components/directories/detail/members/MembersPage';
 import { canManageMembers } from '@/lib/permissions';
 import { notFound } from 'next/navigation';
+
+export async function generateMetadata(): Promise<Metadata> {
+    const t = await getTranslations('metadata.pages');
+    return { title: t('members') };
+}
 
 type Params = { params: Promise<{ id: string }> };
